app/requirements.txt
pandas
psycopg2-binary
requests
beautifulsoup4